doit: Split out cmd_ilpc
