+++ Makefile
@@ -22,9 +22,11 @@ endif
 LDFLAGS = -flto
 
 SRCS += uart/suart.c uart/mux.c uart/vuart.c
-SRCS += prompt.c ast.c ahb.c p2a.c shell.c pci.c l2a.c sio.c ilpc.c doit.c
+SRCS += prompt.c ast.c ahb.c p2a.c shell.c pci.c l2a.c sio.c ilpc.c
 SRCS += clk.c wdt.c sfc.c flash.c mmio.c devmem.c log.c priv.c debug.c rev.c
 SRCS += ts16.c tty.c otp.c
+SRCS += cmd/ilpc.c
+SRCS += doit.c
 
 ifneq (,$(wildcard arch/$(HOST)/lpc.c))
 CFLAGS += -DHAVE_LPC=1
