+++ doit.c
@@ -45,6 +45,8 @@
 #define DUMP_RAM_WIN  (8 << 20)
 #define SFC_FLASH_WIN (64 << 10)
 
+int cmd_ilpc(const char *name, int argc, char *argv[]);
+
 static void help(const char *name)
 {
     printf("%s: " VERSION "\n", name);
@@ -72,43 +74,6 @@ static void help(const char *name)
     printf("%s otp read strap [INTERFACE [IP PORT USERNAME PASSWORD]]\n", name);
     printf("%s otp write strap BIT VALUE [INTERFACE [IP PORT USERNAME PASSWORD]]\n", name);
     printf("%s otp write conf WORD BIT [INTERFACE [IP PORT USERNAME PASSWORD]]\n", name);
-}
-
-static int cmd_ilpc(const char *name, int argc, char *argv[])
-{
-    struct ilpcb _ilpcb, *ilpcb = &_ilpcb;
-    struct ahb _ahb, *ahb = &_ahb;
-    int cleanup;
-    int rc;
-
-    rc = ilpcb_init(ilpcb);
-    if (rc < 0) {
-        bool denied = (rc == -EACCES || rc == -EPERM);
-        if (denied && !priv_am_root()) {
-            priv_print_unprivileged(name);
-        } else {
-            errno = -rc;
-            perror("ilpcb_init");
-        }
-        exit(EXIT_FAILURE);
-    }
-
-    ahb_use(ahb, ahb_ilpcb, ilpcb);
-    rc = ast_ahb_access(name, argc, argv, ahb);
-    if (rc) {
-        errno = -rc;
-        perror("ast_ahb_access");
-        exit(EXIT_FAILURE);
-    }
-
-    cleanup = ilpcb_destroy(ilpcb);
-    if (cleanup) {
-        errno = -cleanup;
-        perror("ilpcb_destroy");
-        exit(EXIT_FAILURE);
-    }
-
-    return 0;
 }
 
 static int cmd_p2a(const char *name, int argc, char *argv[])
